Update Makefile with proper dependency-graph
Dependencies are implemented in the form of empty targets because no
intermediate files are generated that can carry the dependencies properly. The
touching of files when the dependency chains update cause some
unnecessary building of module files and can sometimes be a little annoying in
vim but are nevertheless preferred for now.