Fix PreR operator when pontryagin contains empty polytope
Computation of the PreR operator assumed that the array returned by
union.pontryagin is empty when the Pontryagin difference is empty. It is
however possible that the array contains an empty polytope. In that case the
PreR operator returned a wrong result.
Remove the empty polytopes to obtain correct results.