Remove analysis worker
Because the analysis takes so little time compared to the game construction,
the handling of an additional worker doesn't really bring many benefits.
Integrating the analysis into the system worker simplifies the code and makes
it easier to provide meaningful analysis statistics to the inspector.

This partially reverts changes from 1315b06, a2b6c57 and 399cb7b.